made changes based on suggestions
