+++ tooling-server-configuration/src/main/java/org/jboss/eap/qe/microprofile/tooling/server/logwatch/LogReader.java
@@ -5,13 +5,13 @@ import java.util.regex.Pattern;
 /**
  * Interface describing log watcher which checks log for a pattern.
  */
-public interface LogWatcher {
+public interface LogReader {
 
     /**
-     * Perform search in log or its excerpt and find if a line matches the pattern.
+     * Perform search in log or its excerpt and find if a line that matches the pattern.
      * @param pattern a pattern which will be the log line matched against
      * @return true if log line matching pattern was found in log, false otherwise
      */
-    boolean wasLineWithPatternLogged(final Pattern pattern);
+    boolean wasLineLogged(final Pattern pattern);
 
 }
